Update to fireRound function
fireRound now ensures that a target is selected so that it doesn't cause an error if a fire button is pressed before a target is selected. Also, the checkWin function is integrated into the fireRound function now, but for now it just prints to the console if a player wins.